Add files via upload
If user type total tips of the day. 70% of tip goes to waitress and 30% of tip goes to kitchen(Chef, dishwasher...). Then waitress will divide their tips depends on how much time they work during the day. For example, if a restaurant make $1000 tips then tips will be divided $700 to waitress and $300 to kitchen. Four waitress work during a day and each waitress work 3hr, 4hr, 3hr, and 2hr. Then each waitress will make $175.00, $233.33, $175.00, and $116.67.
This program is not done yet. It has to be added more details and GUI